requirements.txt
python-telegram-bot==20.3
reportlab==4.0.4
pandas==2.0.3